Optimize performance by minimizing number of subscriptions and parallel fetches
Supabase realtime has an undocumented limitation that the subscription
channel name must be unique. https://github.com/supabase/realtime/issues/271
It is very badly documented, but the channel name can actually be any
string. So we can fix this by using the store name for the subscription
channel.

Fixing this resulted in more distinct subscriptions, so I had to put some
effort into reducing this number back down to the minimum.
If a store is already subscribed, we don't have to subscribe again. This
was especially important for the abteilungen store, which is
re-subscribed every time the group scores change (I think because of the
group ranking).

Also, force reloading should still not cause overlapping duplicate
requests. The forceReload flag should just affect the case when there is
already data present in the store.