Enable all users to read the profiles
This was added as an emergency measure during the game of 2023. We
noticed that operators could not read the names of the other operators,
specifically in the author names displayed on chat messages. Likely,
players couldn't see all chat message author names either.
The reason was probably that non-admins did not have read access to all
the profiles. In the heat of the moment, I just activated reading all
other profiles for authenticated users. Maybe there is a better way to
handle this, e.g. allow reading all group partners as well as all
operators and admins, not just the abteilung's assigned operator, at all
times.